✨ admin can now delete all quizzes

+++ pages/admin/index.vue
@@ -1,17 +1,43 @@
 <script setup lang="ts">
 const user = useUser()
+const info = ref('')
 
 if (!user.value || !user.value.isAdmin) {
     navigateTo('/')
 }
 
-function clearUser() {
-    useFetch('/api/ADMIN/clearUsers ', {
+async function clearUser() {
+    if (!confirm('Are you sure you want to clear all users?')) {
+        return
+    }
+    await useFetch('/api/ADMIN/clearUsers ', {
         method: 'delete',
         headers: {
             Authorization: `Bearer ${user.value.token}`
         }
     })
+    info.value = 'Users cleared'
+    clearInfoIn(3000)
+}
+
+async function clearQuiz() {
+    if (!confirm('Are you sure you want to clear the quiz?')) {
+        return
+    }
+    await useFetch('/api/ADMIN/clearQuiz ', {
+        method: 'delete',
+        headers: {
+            Authorization: `Bearer ${user.value.token}`
+        }
+    })
+    info.value = 'Quiz cleared'
+    clearInfoIn(3000)
+}
+
+function clearInfoIn(ms: number) {
+    setTimeout(() => {
+        info.value = ''
+    }, ms)
 }
 </script>
 
@@ -26,8 +52,13 @@ function clearUser() {
                 Options
             </h2>
 
+            <p class="info">
+                {{ info }}
+            </p>
+
             <admin-option
                 option-text="Clear Quiz"
+                @click="clearQuiz"
             />
 
             <admin-option
@@ -66,6 +97,12 @@ h1 {
 h2 {
     font-size: 2rem;
     margin-bottom: 1rem;
+}
+
+.info {
+    font-size: large;
+    margin-bottom: 1rem;
+    color: green;
 }
 
 .option-users {
